+++ storage/store/full_kv.go
@@ -3,6 +3,7 @@ package store
 import (
 	"context"
 	"fmt"
+	"io"
 
 	pbssinternal "github.com/streamingfast/substreams/pb/sf/substreams/intern/v2"
 	"github.com/streamingfast/substreams/storage/store/marshaller"
@@ -34,6 +35,65 @@ func (s *FullKV) DerivePartialStore(initialBlock uint64) *PartialKV {
 		initialBlock: initialBlock,
 		seen:         make(map[string]bool),
 	}
+}
+
+var ErrNoQuickSaveStore = fmt.Errorf("no quick save store")
+
+func (s *FullKV) QuickLoad(ctx context.Context, atBlockHash string) error {
+	if s.quickSaveStore == nil {
+		return ErrNoQuickSaveStore
+	}
+
+	filename := atBlockHash + ".quicksave"
+	s.logger.Debug("loading full store state from temporary file", zap.String("fileName", filename), zap.String("module_hash", s.moduleHash))
+
+	r, err := s.quickSaveStore.OpenObject(ctx, filename)
+	if err != nil {
+		return fmt.Errorf("opening file %q (in module %q): %w", filename, s.moduleHash, err)
+	}
+
+	defer r.Close()
+	data, err := io.ReadAll(r)
+	if err != nil {
+		return fmt.Errorf("reading data: %w", err)
+	}
+
+	storeData, size, err := s.marshaller.Unmarshal(data)
+	if err != nil {
+		return fmt.Errorf("unmarshal store: %w", err)
+	}
+
+	s.kv = storeData.Kv
+	s.totalSizeBytes = size
+	if s.kv == nil {
+		s.kv = make(map[string][]byte)
+	}
+
+	s.logger.Debug("full store loaded", zap.String("fileName", filename), zap.Int("key_count", len(s.kv)), zap.Uint64("data_size", size))
+	return nil
+}
+
+func (s *FullKV) QuickSave(ctx context.Context, atBlockHash string) error {
+	if s.quickSaveStore == nil {
+		return ErrNoQuickSaveStore
+	}
+	s.logger.Debug("writing temporary store state", zap.Object("store", s))
+
+	stateData := &marshaller.StoreData{
+		Kv: s.kv,
+	}
+
+	content, err := s.marshaller.Marshal(stateData)
+	if err != nil {
+		return fmt.Errorf("marshal kv state: %w", err)
+	}
+
+	fw := &fileWriter{
+		store:    s.quickSaveStore,
+		filename: atBlockHash + ".quicksave",
+		content:  content,
+	}
+	return fw.Write(ctx)
 }
 
 func (s *FullKV) Load(ctx context.Context, file *FileInfo) error {