+++ tools/analytics_store_stats.go
@@ -115,6 +115,7 @@ func StoreStatsE(cmd *cobra.Command, args []string) error {
 				module.GetKind().(*pbsubstreams.Module_KindStore_).KindStore.UpdatePolicy,
 				module.GetKind().(*pbsubstreams.Module_KindStore_).KindStore.ValueType,
 				baseDStore,
+				nil,
 			)
 			if err != nil {
 				zlog.Error("creating store config", zap.Error(err))