+++ docs/release-notes/change-log.md
@@ -11,7 +11,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-### Server engine
+### Server
+
+* Implement "QuickSave" feature to save the state of "live running" substreams stores when shutting down, and then resume processing from that point if the cursor matches.
+  - enabled if the "QuickSaveStoreURL" attribute is not empty in the tier1 config
+  - requires the "CheckPendingShutdown" module to be passed to the app via NewTier1()
 
 * Rust modules will now be executed with `wasmtime` by default instead of `wazero`.
   - Prevents the whole server from stalling in certain memory-intensive operations in wazero.